works clean. The project need XML and SMALL edit

+++ WastewaterTreatment/SettlingTank.cs
@@ -8,9 +8,9 @@ namespace WastewaterTreatment
 {
     internal class SettlingTank
     {
-        public int Sedimentation(int microorganism, AerationTank aeration)
+        public double Sedimentation(int microorganism, AerationTank aeration)
         {
-            int sewageSludgeAmount = 0;
+            double sewageSludgeAmount = 0;
             if(microorganism > 100)
             {
                 sewageSludgeAmount = microorganism - 100;
