+++ WastewaterTreatment/Treatment.cs
@@ -9,7 +9,7 @@ namespace WastewaterTreatment
     abstract class Treatment
     {
         public int efficiency { get; private set; }
-        public int removedAmount { get; protected set; }
+        public double removedAmount { get; protected set; }
         public Treatment(int efficiency)
         {
             this.efficiency = efficiency;
